Added scheduled spawn teleport
